Add a base for the rest api
Also includes a new component for the indicator feed data.  I don't completely like this, so still WIP, but trying to keep this simple to bootstrap and defer the introduction of an actual database component till needed.

This commit also introduces a way to introduce test data for one of the interfaced components when testing the base.  I am not sure that it isn't better for the compromise-indicator component to directly reference a feed component, but I like the flexibility of passing data in while testing at this point.