Missed a reference to a build configuration (noted elsewhere I think)
This config had been copied from the clj-poly tool project examples folder but
was just what was needed out of the box when using tools.build. It seems like
there should be scaffolding option when creating a Polylith workspace to
bootstrap such a configuration for any projects that are created.